// src/mutation-types.ts
export enum MutationType {
  ADD_PACKAGE_TO_TRANSFER_PACKAGE_LIST = "ADD_PACKAGE_TO_TRANSFER_PACKAGE_LIST",
  DEQUEUE_TASK = "DEQUEUE_TASK",
  ENQUEUE_TASK = "ENQUEUE_TASK",
  PURGE_TASK_QUEUE = "PURGE_TASK_QUEUE",
  REMOVE_PACKAGE_FROM_TRANSFER_PACKAGE_LIST = "REMOVE_PACKAGE_FROM_TRANSFER_PACKAGE_LIST",
  CLEAR_TRANSFER_PACKAGE_LIST = "CLEAR_TRANSFER_PACKAGE_LIST",
  SET_TRANSFER_PACKAGE_LIST = "SET_TRANSFER_PACKAGE_LIST",
  RESET_STATE = "RESET_STATE",
  RESET_TRACKED_INTERACTIONS = "RESET_TRACKED_INTERACTIONS",
  SELECT_VIEW = "SELECT_VIEW",
  SET_ACCOUNT_ENABLED = "SET_ACCOUNT_ENABLED",
  SET_AUTH_STATE = "SET_AUTH_STATE",
  SET_BUILDER_MODAL_DISPLAY_STATE = "SET_BUILDER_MODAL_DISPLAY_STATE",
  SET_CONTACT_DATA = "SET_CONTACT_DATA",
  SET_CURRENT_VERSION = "SET_CURRENT_VERSION",
  SET_DEBUG_MODE = "SET_DEBUG_MODE",
  SET_DEMO_MODE = "SET_DEMO_MODE",
  SET_ERROR_MESSAGE = "SET_ERROR_MESSAGE",
  SET_EXPANDED_OVERLAY = "SET_EXPANDED_OVERLAY",
  SET_FINALIZE_SALES_RECEIPTS_CONSECUTIVE_ERROR_TOTAL = "SET_FINALIZE_SALES_RECEIPTS_CONSECUTIVE_ERROR_TOTAL",
  SET_FINALIZE_SALES_RECEIPTS_LICENSE = "SET_FINALIZE_SALES_RECEIPTS_LICENSE",
  SET_FINALIZE_SALES_RECEIPTS_READOUT = "SET_FINALIZE_SALES_RECEIPTS_READOUT",
  SET_FINALIZE_SALES_RECEIPTS_RUNNING_TOTAL = "SET_FINALIZE_SALES_RECEIPTS_RUNNING_TOTAL",
  SET_FINALIZE_SALES_RECEIPTS_STATE = "SET_FINALIZE_SALES_RECEIPTS_STATE",
  SET_FINALIZE_SALES_RECEIPTS_STOP_DATE = "SET_FINALIZE_SALES_RECEIPTS_STOP_DATE",
  SET_FLASH_MESSAGE = "SET_FLASH_MESSAGE",
  SET_LOADING_MESSAGE = "SET_LOADING_MESSAGE",
  SET_MOCK_DATA_MODE = "SET_MOCK_DATA_MODE",
  SET_MUTE_ANALYTICS = "SET_MUTE_ANALYTICS",
  SET_OMNI_QUERY_STRING = "SET_OMNI_QUERY_STRING",
  // SET_PACKAGE_QUERY_STRING = 'SET_PACKAGE_QUERY_STRING',
  // SET_PACKAGE_SEARCH_FILTERS = 'SET_PACKAGE_SEARCH_FILTERS',
  SET_REDIRECT = "SET_REDIRECT",
  SET_SEARCH_MODAL_VIEW = "SET_SEARCH_MODAL_VIEW",
  // SET_SHOW_PACKAGE_SEARCH_RESULTS = 'SET_SHOW_PACKAGE_SEARCH_RESULTS',
  SET_SHOW_TRANSFER_SEARCH_RESULTS = "SET_SHOW_TRANSFER_SEARCH_RESULTS",
  SET_TAG_QUERY_STRING = "SET_TAG_QUERY_STRING",
  SET_TAG_SEARCH_FILTERS = "SET_TAG_SEARCH_FILTERS",
  SET_TRANSFER_SEARCH_FILTERS = "SET_TRANSFER_SEARCH_FILTERS",
  SET_VOID_TAGS_CONSECUTIVE_ERROR_TOTAL = "SET_VOID_TAGS_CONSECUTIVE_ERROR_TOTAL",
  SET_VOID_TAGS_DATA = "SET_VOID_TAGS_DATA",
  SET_VOID_TAGS_LICENSE = "SET_VOID_TAGS_LICENSE",
  SET_VOID_TAGS_READOUT = "SET_VOID_TAGS_READOUT",
  SET_VOID_TAGS_RUNNING_TOTAL = "SET_VOID_TAGS_RUNNING_TOTAL",
  SET_VOID_TAGS_STATE = "SET_VOID_TAGS_STATE",
  TOGGLE_DEBUG_MODE = "TOGGLE_DEBUG_MODE",
  TOGGLE_DEMO_MODE = "TOGGLE_DEMO_MODE",
  TOGGLE_EXPANDED_OVERLAY = "TOGGLE_EXPANDED_OVERLAY",
  TOGGLE_PAUSE_TASK_QUEUE = "TOGGLE_PAUSE_TASK_QUEUE",
  UPDATE_ACCOUNT_SETTINGS = "UPDATE_ACCOUNT_SETTINGS",
  UPDATE_CREDENTIALS = "UPDATE_CREDENTIALS",
  UPDATE_METRC_STATUS_DATA = "UPDATE_METRC_STATUS_DATA",
  UPDATE_OVERLAY_VIEW = "UPDATE_OVERLAY_VIEW",
  UPDATE_TRACKED_INTERACTIONS = "UPDATE_TRACKED_INTERACTIONS",
  // Deprecated
  REMOVE_PACKAGE_FROM_TRANSFER_TAG_LIST = "REMOVE_PACKAGE_FROM_TRANSFER_TAG_LIST",
  ADD_PACKAGE_TO_TRANSFER_TAG_LIST = "ADD_PACKAGE_TO_TRANSFER_TAG_LIST",
}